Rearrange code structure
