+++ src/gameboy/cartridge/mbc1.rs
@@ -1,4 +1,5 @@
-use crate::{conv, mmu::MemoryMapped};
+use super::super::mmu::MemoryMapped;
+use crate::conv;
 
 use super::{
     cartridge::Cartridge, cartridge_header::CartridgeHeader, cartridge_type::CartridgeType,
