+++ src/gameboy/instructions.rs
@@ -1,5 +1,5 @@
-use crate::mmu::{IE_REG, IF_REG, MMU};
-use crate::registers::Registers;
+use super::mmu::{IE_REG, IF_REG, MMU};
+use super::registers::Registers;
 
 pub fn _op_cycles(op: u8) -> u32 {
     const OP_CYCLES: [u32; 256] = [
